# docs/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "pwnden local player API",
    "version": "1.4.0",
    "description": "Local player API implemented by the Go server and serve command."
  },
  "servers": [{"url": "/api/v1"}],
  "security": [{"LocalSession": []}],
  "paths": {
...
    "/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "summary": "List this server's retained problem environments",
        "responses": {
          "200": {"description": "Retained environments and policy", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WorkspaceList"}}}},
          "default": {"$ref": "#/components/responses/Failure"}
        }
      }
    },
...
    "/problems": {
      "get": {
        "operationId": "listProblems",
        "summary": "List installed problems in slug order",
        "responses": {
          "200": {
            "description": "Available problem summaries",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProblemList"}}}
          },
          "default": {"$ref": "#/components/responses/Failure"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "LocalSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Random token for one local server process; separate from challenge flags. Loopback binding and Host/Origin checks are also required."
      }
    },
...
    "responses": {
      "Failure": {
        "description": "Public error. Application status mappings and transport failures are defined in web-api.md. Raw causes, Docker output, host paths and recovered flags remain private.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorResponse"}}}
      }
    },
    "schemas": {
      "Slug": {"type": "string", "minLength": 1, "maxLength": 40, "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"},
      "Kind": {"type": "string", "enum": ["file", "service"]},
...
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["slug", "title", "category", "kind"],
        "properties": {
          "slug": {"$ref": "#/components/schemas/Slug"},
          "title": {"type": "string", "minLength": 1},
          "category": {"type": "string"},
          "kind": {"$ref": "#/components/schemas/Kind"}
        }
      },
      "WorkspaceList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["workspaces", "limit", "idle_seconds"],
        "properties": {
          "limit": {"type": "integer", "const": 10},
          "idle_seconds": {"type": "integer", "const": 600},
          "workspaces": {
            "type": "array", "maxItems": 10,
            "items": {
              "type": "object", "additionalProperties": false,
              "required": ["slug", "title", "connected", "expires_at"],
              "properties": {
                "slug": {"$ref": "#/components/schemas/Slug"},
                "title": {"type": "string"},
                "connected": {"type": "boolean"},
                "expires_at": {"type": ["string", "null"], "format": "date-time"}
              }
            }
          }
        }
      },
      "ProblemList": {
        "type": "object",
        "additionalProperties": false,
        "required": ["problems"],
        "properties": {"problems": {"type": "array", "items": {"$ref": "#/components/schemas/Problem"}}}
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "enum": ["invalid_argument", "not_found", "already_running", "not_running", "workspace_full", "terminal_busy", "workspace_busy", "incompatible_contract", "validation_failed", "setup_required", "setup_failed", "canceled", "deadline_exceeded", "execution_failed", "cleanup_failed", "internal_error", "invalid_request", "unauthorized", "forbidden", "method_not_allowed", "payload_too_large", "unsupported_media_type"]
              },
              "message": {"type": "string"},
              "operation": {"type": "string"},
              "slug": {"type": "string"}
            }
          }
        }
      }
    }
  }
}